requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
tensorflow==2.6.1
pillow==8.0.1
python-multipart==0.0.5
pytest==6.1.2